app/requirements.txt
aiohttp==3.9.5

aiosignal==1.3.1

altair==4.0.0

anyio==4.4.0

asgiref==3.8.1

attrs==23.2.0

backoff==2.2.1

bcrypt==4.1.3

blinker==1.8.2

build==1.2.1

cachetools==5.3.3

certifi==2024.6.2

charset-normalizer==3.3.2

chroma-hnswlib==0.7.5

chromadb==0.5.4

click==8.1.7

colorama==0.4.6

coloredlogs==15.0.1

dataclasses-json==0.5.14

Deprecated==1.2.14

dnspython==2.6.1

email_validator==2.2.0

entrypoints==0.4

faiss-cpu==1.7.4

fastapi==0.111.1

fastapi-cli==0.0.4

filelock==3.15.4

flatbuffers==24.3.25

frozenlist==1.4.1

fsspec==2024.6.1

gitdb==4.0.11

GitPython==3.1.43

google-auth==2.32.0

googleapis-common-protos==1.63.2

greenlet==3.0.3

grpcio==1.65.1

h11==0.14.0

httpcore==1.0.5

httptools==0.6.1

httpx==0.27.0

huggingface-hub==0.24.0

humanfriendly==10.0

idna==3.7

importlib_metadata==7.1.0

importlib_resources==6.4.0

Jinja2==3.1.4

jsonpatch==1.33

jsonpointer==3.0.0

jsonschema==4.22.0

jsonschema-specifications==2023.12.1

kubernetes==30.1.0

langchain==0.2.5

langchain-chroma==0.1.2

langchain-community==0.2.5

langchain-core==0.2.22

langchain-text-splitters==0.2.2

langdetect==1.0.9

langsmith==0.1.82

lxml==5.2.2

markdown-it-py==3.0.0

MarkupSafe==2.1.5

marshmallow==3.21.3

mdurl==0.1.2

mmh3==4.1.0

monotonic==1.6

mpmath==1.3.0

multidict==6.0.5

mypy-extensions==1.0.0

numexpr==2.10.1

numpy==1.26.4

oauthlib==3.2.2

onnxruntime==1.18.1

openai==0.27.6

openapi-schema-pydantic==1.2.4

opentelemetry-api==1.25.0

opentelemetry-exporter-otlp-proto-common==1.25.0

opentelemetry-exporter-otlp-proto-grpc==1.25.0

opentelemetry-instrumentation==0.46b0

opentelemetry-instrumentation-asgi==0.46b0

opentelemetry-instrumentation-fastapi==0.46b0

opentelemetry-proto==1.25.0

opentelemetry-sdk==1.25.0

opentelemetry-semantic-conventions==0.46b0

opentelemetry-util-http==0.46b0

orjson==3.10.5

overrides==7.7.0

packaging==24.1

pandas==2.2.2

pillow==10.3.0

posthog==3.5.0

protobuf==3.20.3

pyarrow==16.1.0

pyasn1==0.6.0

pyasn1_modules==0.4.0

pydantic==1.10.17

pydeck==0.9.1

Pygments==2.18.0

Pympler==1.0.1

PyPDF2==3.0.1

PyPika==0.48.9

pyproject_hooks==1.1.0

pyreadline3==3.4.1

python-dateutil==2.9.0.post0

python-docx==1.1.2

python-dotenv==1.0.0

python-multipart==0.0.9

python-pptx==0.6.23

pytz==2024.1

PyYAML==6.0.1

referencing==0.35.1

regex==2024.5.15

requests==2.32.3

requests-oauthlib==2.0.0

rich==13.7.1

rpds-py==0.18.1

rsa==4.9

semver==3.0.2

shellingham==1.5.4

six==1.16.0

smmap==5.0.1

sniffio==1.3.1

SQLAlchemy==2.0.31

starlette==0.37.2

streamlit==1.18.1

sympy==1.13.1

tenacity==8.4.2

tiktoken==0.4.0

tokenizers==0.19.1

toml==0.10.2

toolz==0.12.1

tornado==6.4.1

tqdm==4.66.4

typer==0.12.3

typing-inspect==0.9.0

typing_extensions==4.12.2

tzdata==2024.1

tzlocal==5.2

urllib3==2.2.2

uvicorn==0.30.3

validators==0.28.3

watchdog==4.0.1

watchfiles==0.22.0

websocket-client==1.8.0

websockets==12.0

wrapt==1.16.0

XlsxWriter==3.2.0

yarl==1.9.4

zipp==3.19.2